Adding once-per-turn check to CanTrigger (BT22-052 & EX9-056)
Adding the same OPT check that CanActivate has to CanTrigger.

Context: The autoProcessor_CutIn processor doesn't update as much as the regular autoProcessor, which regularly cleans its queue of stacked skills. In any scenario where autoProcessor_CutIn .StackSkillInfos adds triggered effects to its list of stacked effects, but those effects can't activate right away (like if the effect is OPT and has already been used), the triggered effect stays in the list and gets stale, until it is cleaned out at the next action.  If that next action happens on the opponent's turn, the OPT check is cleared and the effect can activate.  Adding the check in CanTrigger stops that effect from ever getting added if it's already been used the max number of times that turn.