ESS fixes - BT18_047, BT18_050, BT18_012, BT18_014
Fixing bug "ESS does not trigger if opponent has no Digimon" for multiple cards. Copied implementation from BT22_029, removing check for opponent digimon to allow the effect to activate.

For BT18_047 and BT18_050, added the HasMatchConditionOpponentsPermanent check to the ActivateCoroutine function, to make sure an enemy digimon exists to target.